+++ examples/src/main/java/software/xdev/micromigration/examples/practical/embedded/MainPracticalWithMigrationEmbeddedStorageManager.java
@@ -1,9 +1,8 @@
 package software.xdev.micromigration.examples.practical.embedded;
 
+import software.xdev.micromigration.examples.practical.v0.BusinessBranch;
 import software.xdev.micromigration.microstream.MigrationEmbeddedStorage;
 import software.xdev.micromigration.microstream.MigrationEmbeddedStorageManager;
-import software.xdev.micromigration.examples.practical.v0.BusinessBranch;
-import software.xdev.micromigration.examples.practical.v0.Customer;
 import software.xdev.micromigration.migrater.ExplicitMigrater;
 
 
@@ -19,7 +18,7 @@ import software.xdev.micromigration.migrater.ExplicitMigrater;
  * @author Johannes Rabauer
  *
  */
-public class MainPracticalWithMigrationEmbeddedStorageManager 
+public class MainPracticalWithMigrationEmbeddedStorageManager
 {
 	public static void main(String[] args) 
 	{
@@ -27,7 +26,7 @@ public class MainPracticalWithMigrationEmbeddedStorageManager
 		final ExplicitMigrater emptyMigrater = new ExplicitMigrater();
 		try(MigrationEmbeddedStorageManager storageManager = MigrationEmbeddedStorage.start(emptyMigrater))
 		{
-			storageManager.setRoot(createDummyBranch());
+			storageManager.setRoot(BusinessBranch.createDummyBranch());
 			storageManager.storeRoot();
 			System.out.println(storageManager.root().toString());
 		}
@@ -52,21 +51,5 @@ public class MainPracticalWithMigrationEmbeddedStorageManager
 		}
 	}
 	
-	private static BusinessBranch createDummyBranch()
-	{
-		BusinessBranch branch = new BusinessBranch();
-		Customer customer1 = new Customer();
-		customer1.name   = "Mick Fleetwood";
-		customer1.number = 1;
-		customer1.street = "Fleetwood Street";
-		customer1.city   = "Redruth";
-		branch.customers.add(customer1);
-		Customer customer2 = new Customer();
-		customer2.name   = "Lindsey Buckingham";
-		customer2.number = 2;
-		customer2.street = "Mac Street";
-		customer2.city   = "Palo Alto";
-		branch.customers.add(customer2);
-		return branch;
-	}
+
 }