+++ examples/src/main/java/software/xdev/micromigration/examples/practical/embedded/UpdateToV1_0.java
@@ -1,12 +1,12 @@
 package software.xdev.micromigration.examples.practical.embedded;
 
 import software.xdev.micromigration.examples.practical.v1AndHigher.Address;
+import software.xdev.micromigration.examples.practical.v1AndHigher.BusinessBranch;
+import software.xdev.micromigration.examples.practical.v1AndHigher.Customer;
+import software.xdev.micromigration.microstream.MigrationEmbeddedStorageManager;
 import software.xdev.micromigration.microstream.MigrationScript;
 import software.xdev.micromigration.scripts.Context;
 import software.xdev.micromigration.version.MigrationVersion;
-import software.xdev.micromigration.examples.practical.v1AndHigher.BusinessBranch;
-import software.xdev.micromigration.examples.practical.v1AndHigher.Customer;
-import one.microstream.storage.embedded.types.EmbeddedStorageManager;
 
 
 public class UpdateToV1_0 implements MigrationScript<software.xdev.micromigration.examples.practical.v0.BusinessBranch>
@@ -18,7 +18,7 @@ public class UpdateToV1_0 implements MigrationScript<software.xdev.micromigratio
 	}
 	
 	@Override
-	public void migrate(Context<software.xdev.micromigration.examples.practical.v0.BusinessBranch, EmbeddedStorageManager> context)
+	public void migrate(Context<software.xdev.micromigration.examples.practical.v0.BusinessBranch, MigrationEmbeddedStorageManager> context)
 	{
 		System.out.println("Executing Script for v1.0...");
 		software.xdev.micromigration.examples.practical.v0.BusinessBranch oldBranch = context.getMigratingObject();