+++ examples/src/main/java/software/xdev/micromigration/examples/practical/migrationManager/MainPracticalWithMigrationManager.java
@@ -1,12 +1,11 @@
 package software.xdev.micromigration.examples.practical.migrationManager;
 
-import software.xdev.micromigration.microstream.MigrationManager;
-import software.xdev.micromigration.examples.practical.v0.BusinessBranch;
-import software.xdev.micromigration.examples.practical.v0.Customer;
-import software.xdev.micromigration.migrater.ExplicitMigrater;
-import software.xdev.micromigration.version.VersionedObject;
 import one.microstream.storage.embedded.types.EmbeddedStorage;
 import one.microstream.storage.embedded.types.EmbeddedStorageManager;
+import software.xdev.micromigration.examples.practical.v0.BusinessBranch;
+import software.xdev.micromigration.microstream.MigrationManager;
+import software.xdev.micromigration.migrater.ExplicitMigrater;
+import software.xdev.micromigration.version.VersionedObject;
 
 
 /**
@@ -32,7 +31,7 @@ public class MainPracticalWithMigrationManager
 		//V0.0
 		try(EmbeddedStorageManager storageManager = EmbeddedStorage.start())
 		{
-			VersionedObject<BusinessBranch> versionedBranch = new VersionedObject<>(createDummyBranch());
+			VersionedObject<BusinessBranch> versionedBranch = new VersionedObject<>(BusinessBranch.createDummyBranch());
 			storageManager.setRoot(versionedBranch);
 			storageManager.storeRoot();
 			System.out.println(storageManager.root().toString());
@@ -59,23 +58,5 @@ public class MainPracticalWithMigrationManager
 				.migrate(versionedBranch);
 			System.out.println(storageManager.root().toString());
 		}
-	}
-	
-	private static BusinessBranch createDummyBranch()
-	{
-		BusinessBranch branch = new BusinessBranch();
-		Customer customer1 = new Customer();
-		customer1.name   = "Mick Fleetwood";
-		customer1.number = 1;
-		customer1.street = "Fleetwood Street";
-		customer1.city   = "Redruth";
-		branch.customers.add(customer1);
-		Customer customer2 = new Customer();
-		customer2.name   = "Lindsey Buckingham";
-		customer2.number = 2;
-		customer2.street = "Mac Street";
-		customer2.city   = "Palo Alto";
-		branch.customers.add(customer2);
-		return branch;
 	}
 }