+++ examples/src/main/java/software/xdev/micromigration/examples/practical/migrationManager/UpdateToV1_0.java
@@ -3,11 +3,11 @@ package software.xdev.micromigration.examples.practical.migrationManager;
 import software.xdev.micromigration.examples.practical.v0.BusinessBranch;
 import software.xdev.micromigration.examples.practical.v0.Customer;
 import software.xdev.micromigration.examples.practical.v1AndHigher.Address;
+import software.xdev.micromigration.microstream.MigrationEmbeddedStorageManager;
 import software.xdev.micromigration.microstream.MigrationScript;
 import software.xdev.micromigration.scripts.Context;
 import software.xdev.micromigration.version.MigrationVersion;
 import software.xdev.micromigration.version.VersionedObject;
-import one.microstream.storage.embedded.types.EmbeddedStorageManager;
 
 
 public class UpdateToV1_0 implements MigrationScript<VersionedObject<Object>>
@@ -19,7 +19,7 @@ public class UpdateToV1_0 implements MigrationScript<VersionedObject<Object>>
 	}
 
 	@Override
-	public void migrate(Context<VersionedObject<Object>, EmbeddedStorageManager> context)
+	public void migrate(Context<VersionedObject<Object>, MigrationEmbeddedStorageManager> context)
 	{
 		System.out.println("Executing Script for v1.0...");
 		VersionedObject<Object> versionedBranch = context.getMigratingObject();