+++ examples/src/main/java/software/xdev/micromigration/examples/reflective/MainReflective.java
@@ -1,13 +1,18 @@
 package software.xdev.micromigration.examples.reflective;
 
-import java.util.Date;
-
 import software.xdev.micromigration.microstream.MigrationEmbeddedStorage;
 import software.xdev.micromigration.microstream.MigrationEmbeddedStorageManager;
 import software.xdev.micromigration.migrater.ReflectiveMigrater;
 import software.xdev.micromigration.migrater.ScriptInstantiationException;
 
+import java.util.Date;
 
+
+/**
+ * Shows the usage of the {@link ReflectiveMigrater}. Very simple.
+ *
+ * @author Johannes Rabauer
+ */
 public class MainReflective
 {
 	public static void main(String[] args)