+++ microstream-v7/src/main/java/software/xdev/micromigration/microstream/MigrationScript.java
@@ -1,18 +1,20 @@
 package software.xdev.micromigration.microstream;
 
 import software.xdev.micromigration.scripts.Context;
-import one.microstream.storage.embedded.types.EmbeddedStorageManager;
+import software.xdev.micromigration.scripts.VersionAgnosticMigrationScript;
 
 
 /**
  * Interface for scripts to migrate / update datastores.
  * <p>
  * One script is supposed to bring a datastore from a lower version to the target version.
- * After the {@link software.xdev.micromigration.scripts.MigrationScript#migrate(Context)} method is called,
+ * After the {@link VersionAgnosticMigrationScript#migrate(Context)} method is called,
  * the target version is reached.
- * 
+ * <p>
+ * This is a shorthand for {@link VersionAgnosticMigrationScript} for this specific MicroStream version.
+ *
  * @author Johannes Rabauer
  *
  */
-public interface MigrationScript<T> extends software.xdev.micromigration.scripts.MigrationScript<T, EmbeddedStorageManager>
+public interface MigrationScript<T> extends VersionAgnosticMigrationScript<T, MigrationEmbeddedStorageManager>
 {}