+++ reflection/src/test/java/software/xdev/micromigration/migrater/scripts/abstractReflectiveSuperClass/v1_ValidScript.java
@@ -1,13 +1,13 @@
 package software.xdev.micromigration.migrater.scripts.abstractReflectiveSuperClass;
 
+import software.xdev.micromigration.microstream.MigrationEmbeddedStorageManager;
 import software.xdev.micromigration.scripts.Context;
-import one.microstream.storage.embedded.types.EmbeddedStorageManager;
 
 
 public class v1_ValidScript extends AbstractScript
 {
 	@Override
-	public void migrate(Context<Object, EmbeddedStorageManager> context)
+	public void migrate(Context<Object, MigrationEmbeddedStorageManager> context)
 	{
 		//Do nothing
 	}